+++ graphdb/janus/src/main/java/org/apache/atlas/repository/graphdb/janus/cassandra/CassandraVertexDataRepository.java
@@ -56,71 +56,6 @@ class CassandraVertexDataRepository implements VertexDataRepository {
         this.objectMapper = objectMapper;
     }
 
-    /**
-     * Fetches vertex JSON data as strings.
-     */
-    @Override
-    public Map<String, String> fetchVerticesJsonData(List<String> vertexIds) throws AtlasBaseException {
-        if (vertexIds == null || vertexIds.isEmpty()) {
-            return Collections.emptyMap();
-        }
-
-        // Filter out blank IDs
-        List<String> sanitizedIds = new ArrayList<>();
-        for (String id : vertexIds) {
-            if (StringUtils.isNotBlank(id)) {
-                sanitizedIds.add(id);
-            }
-        }
-
-        if (sanitizedIds.isEmpty()) {
-            return Collections.emptyMap();
-        }
-
-        // Split large batches into smaller ones to avoid Cassandra limitations
-        if (sanitizedIds.size() > MAX_IN_CLAUSE_ITEMS) {
-            return fetchLargeBatch(sanitizedIds);
-        } else {
-            return fetchSingleBatch(sanitizedIds);
-        }
-    }
-
-    /**
-     * Fetches vertex data as parsed JsonNodes instead of raw strings.
-     * This is more efficient when the caller needs to work with the JSON directly.
-     *
-     * @param vertexIds List of vertex IDs to fetch
-     * @return Map of vertex ID to parsed JsonNode
-     */
-    @Override
-    public Map<String, JsonNode> fetchVerticesAsJsonNodes(List<String> vertexIds) throws AtlasBaseException {
-        Map<String, String> vertexJsonDataLookup = fetchVerticesJsonData(vertexIds);
-        Map<String, JsonNode> vertexJsonNodesLookup = new HashMap<>(vertexJsonDataLookup.size());
-
-        for (Map.Entry<String, String> entry : vertexJsonDataLookup.entrySet()) {
-            try {
-                JsonNode node = objectMapper.readTree(entry.getValue());
-                vertexJsonNodesLookup.put(entry.getKey(), node);
-            } catch (JsonProcessingException e) {
-                LOG.warn("Failed to parse JSON for vertex ID {}: {}", entry.getKey(), e.getMessage());
-                throw new AtlasBaseException("Failed to parse JSON for vertex ID: " + entry.getKey(), e);
-            }
-        }
-
-        return vertexJsonNodesLookup;
-    }
-
-    @Override
-    public JsonNode fetchVertexAsJsonNode(String vertexId) throws AtlasBaseException {
-        Map<String, String> data = fetchVerticesJsonData(Collections.singletonList(vertexId));
-        try {
-            return objectMapper.readTree(data.get(vertexId));
-        } catch (JsonProcessingException e) {
-            LOG.warn("Failed to parse JSON for vertex ID {}: {}", vertexId, e.getMessage());
-            throw new AtlasBaseException("Failed to parse JSON for vertex ID: " + vertexId, e);
-        }
-    }
-
     @Override
     public void insertVertices(Map<String, String> serialisedVertices) throws AtlasBaseException {
         StringBuilder batchQuery = new StringBuilder();
@@ -140,73 +75,6 @@ class CassandraVertexDataRepository implements VertexDataRepository {
 
         batchQuery.append("APPLY BATCH;");
         session.execute(batchQuery.toString());
-    }
-
-    /**
-     * Fetches a large batch of vertices by splitting into smaller batches.
-     */
-    private Map<String, String> fetchLargeBatch(List<String> vertexIds) {
-        Map<String, String> results = new HashMap<>();
-        List<List<String>> batches = Lists.partition(vertexIds, MAX_IN_CLAUSE_ITEMS);
-
-        // Process each batch
-        for (List<String> batch : batches) {
-            try {
-                Map<String, String> batchResults = fetchSingleBatch(batch);
-                results.putAll(batchResults);
-            } catch (Exception e) {
-                LOG.error("Error fetching batch of vertex data", e);
-                // Continue with other batches even if one fails
-            }
-        }
-
-        return results;
-    }
-
-    /**
-     * Fetches a single batch of vertices with retry logic.
-     * Parses JSON data into JsonNode trees using Jackson.
-     */
-    private Map<String, String> fetchSingleBatch(List<String> vertexIds) throws AtlasBaseException {
-        try {
-
-            // Get or prepare the statement for this batch size
-            PreparedStatement statement = getPreparedStatementForBatchSize(vertexIds.size());
-
-            // Bind values
-            BoundStatement boundStatement = statement.bind();
-            for (int i = 0; i < vertexIds.size(); i++) {
-                boundStatement = boundStatement.setString(i, vertexIds.get(i));
-            }
-
-            // Set query timeout and other options
-            boundStatement = boundStatement
-                    .setConsistencyLevel(ConsistencyLevel.LOCAL_QUORUM);
-
-            // Execute the query
-            ResultSet resultSet = session.execute(boundStatement);
-
-            // Process the results and parse JSON with Jackson
-            Map<String, String> results = new HashMap<>();
-
-
-            for (Row row : resultSet) {
-                String id = row.getString("id");
-                String jsonData = row.getString("json_data");
-                results.put(id, jsonData);
-            }
-
-            return results;
-
-        } catch (QueryValidationException e) {
-            // These are non-recoverable errors with the query itself
-            LOG.error("Invalid query error fetching vertex data: {}", e.getMessage());
-            throw new AtlasBaseException("Invalid query: " + e.getMessage(), e);
-        } catch (Exception e) {
-            // For unexpected errors
-            LOG.error("Unexpected error fetching vertex data", e);
-            throw new AtlasBaseException("Failed to fetch vertex data: " + e.getMessage(), e);
-        }
     }
 
     /**
@@ -246,4 +114,113 @@ class CassandraVertexDataRepository implements VertexDataRepository {
         int numBuckets = 2 << 5; // 2^5=32
         return (int) (Long.parseLong(vertexId) % numBuckets);
     }
+
+    /**
+     * Fetches vertices directly as DynamicVertex objects without intermediate JSON serialization/deserialization.
+     * This is the most efficient method for retrieving vertices from the database.
+     *
+     * @param vertexIds List of vertex IDs to fetch
+     * @return Map of vertex ID to DynamicVertex object
+     */
+    @Override
+    public Map<String, DynamicVertex> fetchVerticesDirectly(List<String> vertexIds) throws AtlasBaseException {
+        if (vertexIds == null || vertexIds.isEmpty()) {
+            return Collections.emptyMap();
+        }
+
+        // Filter out blank IDs
+        List<String> sanitizedIds = new ArrayList<>();
+        for (String id : vertexIds) {
+            if (StringUtils.isNotBlank(id)) {
+                sanitizedIds.add(id);
+            }
+        }
+
+        if (sanitizedIds.isEmpty()) {
+            return Collections.emptyMap();
+        }
+
+        Map<String, DynamicVertex> results = new HashMap<>();
+
+        // Split large batches into smaller ones to avoid Cassandra limitations
+        if (sanitizedIds.size() > MAX_IN_CLAUSE_ITEMS) {
+            List<List<String>> batches = Lists.partition(sanitizedIds, MAX_IN_CLAUSE_ITEMS);
+
+            // Process each batch
+            for (List<String> batch : batches) {
+                try {
+                    Map<String, DynamicVertex> batchResults = fetchSingleBatchDirectly(batch);
+                    results.putAll(batchResults);
+                } catch (Exception e) {
+                    LOG.error("Error fetching batch of vertex data directly", e);
+                    // Continue with other batches even if one fails
+                }
+            }
+        } else {
+            results = fetchSingleBatchDirectly(sanitizedIds);
+        }
+
+        return results;
+    }
+
+    /**
+     * Fetches a single batch of vertices directly as DynamicVertex objects.
+     * Avoids the JSON string parsing overhead.
+     */
+    private Map<String, DynamicVertex> fetchSingleBatchDirectly(List<String> vertexIds) throws AtlasBaseException {
+        try {
+            // Get or prepare the statement for this batch size
+            PreparedStatement statement = getPreparedStatementForBatchSize(vertexIds.size());
+
+            // Bind values
+            BoundStatement boundStatement = statement.bind();
+            for (int i = 0; i < vertexIds.size(); i++) {
+                boundStatement = boundStatement.setString(i, vertexIds.get(i));
+            }
+
+            // Set query timeout and other options
+            boundStatement = boundStatement.setConsistencyLevel(ConsistencyLevel.LOCAL_QUORUM);
+
+            // Execute the query
+            ResultSet resultSet = session.execute(boundStatement);
+
+            // Process the results directly into DynamicVertex objects
+            Map<String, DynamicVertex> results = new HashMap<>(vertexIds.size());
+
+            for (Row row : resultSet) {
+                String id = row.getString("id");
+                String jsonData = row.getString("json_data");
+
+                try {
+                    // Parse the JSON directly to a Map
+                    Map<String, Object> props = objectMapper.readValue(jsonData, Map.class);
+
+                    // Create a DynamicVertex with all properties from the Map
+                    DynamicVertex vertex = new DynamicVertex(props);
+
+                    // Ensure ID is set in the vertex properties
+                    if (!vertex.hasProperty("id")) {
+                        vertex.setProperty("id", id);
+                    }
+
+                    results.put(id, vertex);
+                } catch (Exception e) {
+                    LOG.warn("Failed to convert JSON to DynamicVertex for ID {}: {}", id, e.getMessage());
+                    // Skip this vertex but continue processing others
+                }
+            }
+
+            return results;
+
+        } catch (QueryValidationException e) {
+            // These are non-recoverable errors with the query itself
+            LOG.error("Invalid query error fetching vertex data directly: {}", e.getMessage());
+            throw new AtlasBaseException("Invalid query: " + e.getMessage(), e);
+        } catch (Exception e) {
+            // For unexpected errors
+            LOG.error("Unexpected error fetching vertex data directly", e);
+            throw new AtlasBaseException("Failed to fetch vertex data directly: " + e.getMessage(), e);
+        }
+    }
+
 }
