--- repository/src/main/java/org/apache/atlas/repository/cassandra/Validation.java
@@ -1,27 +1,0 @@
-package org.apache.atlas.repository.cassandra;
-
-import com.datastax.oss.driver.api.core.CqlSession;
-
-import java.util.Arrays;
-import java.util.Map;
-
-public class Validation  {
-
-    public static void main(String[] args)  {
-        CqlSession cqlSession = null;
-        try {
-            CassandraConfig config = new CassandraConfig();
-            cqlSession = config.cqlSession();
-            VertexRetrievalService batchVertexRetrievalService = new VertexRetrievalService(cqlSession);
-            Map<String, DynamicVertex> vertexPropertiesMap = batchVertexRetrievalService.retrieveVertices(Arrays.asList("40964232", "122884120"));
-            System.out.println("vertexPropertiesMap = " + vertexPropertiesMap);
-
-        } catch (Exception e) {
-            System.out.println("Error while closing cqlSession: " + e.getMessage());
-        } finally {
-            if (cqlSession != null) {
-                cqlSession.close();
-            }
-        }
-    }
-}
