--- repository/src/main/java/org/apache/atlas/repository/cassandra/VertexRetrievalService.java
@@ -1,74 +1,0 @@
-package org.apache.atlas.repository.cassandra;
-
-import com.datastax.oss.driver.api.core.CqlSession;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.atlas.AtlasConfiguration;
-import org.apache.atlas.exception.AtlasBaseException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Service;
-
-import javax.inject.Inject;
-import java.util.*;
-
-/**
- * Main entry point for the batch vertex retrieval system.
- * Coordinates the retrieval process and delegates to specialized components.
- */
-
-@Service
-public class VertexRetrievalService {
-    private static final Logger LOG = LoggerFactory.getLogger(VertexRetrievalService.class);
-
-    private final VertexDataRepository repository;
-    private final int defaultBatchSize;
-
-    /**
-     * Creates a new BatchVertexRetrievalService with custom configuration.
-     *
-     * @param session The Cassandra session
-     */
-    @Inject
-    public VertexRetrievalService(CqlSession session) {
-        ObjectMapper objectMapper = new ObjectMapper();
-        this.repository = new CassandraVertexDataRepository(session, objectMapper, AtlasConfiguration.ATLAS_CASSANDRA_VANILLA_KEYSPACE.getString(),
-                AtlasConfiguration.ATLAS_CASSANDRA_VERTEX_TABLE.getString());
-        this.defaultBatchSize = AtlasConfiguration.ATLAS_CASSANDRA_BATCH_SIZE.getInt();
-    }
-
-    /**
-     * Retrieves multiple vertices by their IDs.
-     *
-     * @param vertexIds The list of vertex IDs to retrieve
-     * @return A map of vertex ID to dynamic vertex data
-     */
-    public Map<String, DynamicVertex> retrieveVertices(List<String> vertexIds) throws AtlasBaseException {
-        return retrieveVertices(vertexIds, defaultBatchSize);
-    }
-
-    /**
-     * Retrieves multiple vertices by their IDs with custom batch size.
-     *
-     * @param vertexIds The list of vertex IDs to retrieve
-     * @param batchSize The batch size to use
-     * @return A map of vertex ID to dynamic vertex data
-     */
-    private Map<String, DynamicVertex> retrieveVertices(List<String> vertexIds, int batchSize) throws AtlasBaseException {
-        if (vertexIds == null || vertexIds.isEmpty()) {
-            return Collections.emptyMap();
-        }
-
-        Map<String, DynamicVertex> results = new HashMap<>();
-
-        for (int i = 0; i < vertexIds.size(); i += batchSize) {
-            int endIndex = Math.min(i + batchSize, vertexIds.size());
-            List<String> batch = vertexIds.subList(i, endIndex);
-
-            // Use direct loading approach
-            Map<String, DynamicVertex> batchResults = repository.fetchVerticesDirectly(batch);
-            results.putAll(batchResults);
-        }
-
-        return results;
-    }
-}
